Add login_required decorator for user not accessing to any administration part

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
@@ -9,7 +10,6 @@ from .forms import ProductForm
 def all_products(request):
     """A view to show all products,
     including sorting and searching queries"""
-
     products = Product.objects.all()
     query = None
     categories = None
@@ -71,8 +71,13 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """ Add a product to the store as an admin """
+    if not request.user.is_superuser:
+        messages.error(request, 'Something went wrong, only store owners can access to this.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -92,8 +97,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit an existent product as an admin """
+    if not request.user.is_superuser:
+        messages.error(request, 'Something went wrong, only store owners can access to this.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -116,8 +126,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ Delete product from the store as admin """
+    if not request.user.is_superuser:
+        messages.error(request, 'Something went wrong, only store owners can access to this.')
+        return redirect(reverse('home'))
+        
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted successfully')
